requirements.txt
requests==2.31.0
pandas==2.1.1
tqdm==4.66.1
openpyxl==3.1.2
Pillow==10.0.1
urllib3==2.0.7
gradio==3.48.0